+++ config/module.config.php
@@ -4,18 +4,25 @@ return array(
     'service_manager' => array(
         'factories' => array(
             'SlmQueue\Job\JobPluginManager'     => 'SlmQueue\Factory\JobPluginManagerFactory',
-            'SlmQueue\Options\WorkerOptions'    => 'SlmQueue\Factory\WorkerOptionsFactory',
+            'SlmQueue\Options\ModuleOptions'    => 'SlmQueue\Factory\ModuleOptionsFactory',
             'SlmQueue\Queue\QueuePluginManager' => 'SlmQueue\Factory\QueuePluginManagerFactory'
         ),
     ),
 
     'slm_queue' => array(
-        /**
-         * Parameters for the worker
-         */
-        'worker' => array(
-            'max_runs'   => 100000,
-            'max_memory' => 100 * 1024 * 1024
+        'options' => array(
+            /**
+             * Parameters for the worker
+             */
+            'worker' => array(
+                'max_runs'   => 100000,
+                'max_memory' => 100 * 1024 * 1024
+            ),
+
+            /**
+             * Queue configuration options
+             */
+            'queues' => array(),
         ),
 
         /**
@@ -27,10 +34,5 @@ return array(
          * Queue manager configuration
          */
         'queue_manager' => array(),
-
-        /**
-         * Queue configuration options
-         */
-        'queues' => array(),
     ),
 );